pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    { name = "Raphael da Silva", email = "raphael.h.silva@protonmail.com" },
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "fastapi[standard] (>=0.124.4,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "aio-pika (>=9.5.8,<10.0.0)",
]

[tool.poetry]
packages = [{ include = "auth_service", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.9,<0.15.0)",
    "testcontainers[postgres] (>=4.13.3,<5.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
]

[tool.ruff]
line-length = 79
extend-exclude = ['alembic']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']


[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev src/auth_service/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'
